refactor: rename utils module
